+++ src/stonepy/models/ApiSaveClientPreferenceRequestDTO.py
@@ -10,7 +10,7 @@ from pydantic import Field
 from stonepy._core.models import RequestModel
 
 if TYPE_CHECKING:
-    from .ClientPreferenceKeyDTO import ClientPreferenceKeyDTO
+    from .RequestClientPreferenceKeyDTO import RequestClientPreferenceKeyDTO
 
 
 class ApiSaveClientPreferenceRequestDTO(RequestModel):
@@ -18,7 +18,9 @@ class ApiSaveClientPreferenceRequestDTO(RequestModel):
 
     client_account_id: int = Field(alias="ClientAccountId")
     """Client account identifier"""
-    client_preference: ClientPreferenceKeyDTO | None = Field(default=None, alias="ClientPreference")
+    client_preference: RequestClientPreferenceKeyDTO | None = Field(
+        default=None, alias="ClientPreference"
+    )
     """The list of client preferences key value pairs to be saved"""